feat: fix logo and clean some code
