+++ package.json
@@ -10,27 +10,28 @@
   },
   "dependencies": {
     "@emailjs/browser": "^3.10.0",
-    "aos": "^2.3.4",
     "@types/node": "20.6.3",
     "@types/react": "18.2.22",
     "@types/react-dom": "18.2.7",
+    "aos": "^2.3.4",
     "autoprefixer": "10.4.15",
     "eslint": "8.49.0",
     "eslint-config-next": "13.5.2",
     "next": "13.5.2",
     "next-themes": "^0.2.1",
     "postcss": "8.4.30",
+    "rc-progress": "^3.4.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.11.0",
-    "tailwindcss": "3.3.3",
-    "typescript": "5.2.2",
-    "rc-progress": "^3.4.1",
     "react-masonry-css": "^1.0.16",
     "react-modal": "^3.16.1",
     "react-slick": "^0.29.0",
     "react-toastify": "^9.1.1",
-    "slick-carousel": "^1.8.1"
+    "sharp": "^0.32.6",
+    "slick-carousel": "^1.8.1",
+    "tailwindcss": "3.3.3",
+    "typescript": "5.2.2"
   },
   "devDependencies": {
     "eslint-config-prettier": "^9.0.0",
