+++ src/components/atoms/index.ts
@@ -1,0 +1,2 @@
+export { default as Logo } from './logo';
+export { default as Seo } from './seo';
