+++ src/components/atoms/logo/index.tsx
@@ -1,0 +1,31 @@
+import * as React from 'react';
+import { LogoProps } from './types';
+
+const SvgComponent = (props: LogoProps) => (
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 815.9 360.89" {...props}>
+    <defs>
+      <style>{'.cls-1{fill:currentColor}.cls-2{fill:currentColor}'}</style>
+    </defs>
+    <g id="Layer_2" data-name="Layer 2">
+      <g id="Layer_1-2" data-name="Layer 1">
+        <path
+          d="M423.28 124c3.9-7.49 7.58-15 11.66-22.32 16.14-28.95 35.74-55.27 60.4-77.56a125.23 125.23 0 0 1 32.42-21.75c3.06-1.39 4.77-.84 6.72 1.56a68.64 68.64 0 0 1 12.57 24c.77 2.58 1.76 3.82 4.41 3.15 2.91-.73 4.37 1 5.36 3.19 6.5 14.23 8.8 29 5.35 44.41-1.47 6.56-5.56 11.4-11.3 15a186.13 186.13 0 0 0-44.11 38.91c-.49.6-.89 1.29-1.63 2.37 3.15.86 5.52-.55 7.93-1 11.71-2.46 23.31-2.63 34.74 1.42a30.88 30.88 0 0 1 3.22 1.29c8.73 4.17 9.48 5.89 6.94 15-.7 2.5-1.94 4.9-1.76 8.17a49.53 49.53 0 0 1 13.56 3.87c2.49-10.34 4.82-20.37 7.33-30.36 7.72-30.73 15.68-61.4 23.17-92.18 2.54-10.45 8.27-18.19 16.82-24.21a84.65 84.65 0 0 1 51.77-15.7 29 29 0 0 1 5.41.6c9.76 2 11 3.75 10.1 13.41-.39 4.08-.75 8.17-1.15 12.6 3.34.29 6.24.57 9.14.78 3.39.24 4.8 1.46 4.34 5.35-1.74 14.79-1.63 29.7-2.21 44.56-.05 1.17-.69 2.64.77 3.53 1.39-.13 1.39-1.32 1.69-2.16 5.15-14.64 10.75-29.14 15.23-44 3.39-11.2 10-19.43 19.82-24.94 21.94-12.27 45-14.08 68.94-6.41 3.69 1.19 4.31 3 3.49 6.61-1.2 5.27-2.74 10.54-2.78 16.24 3.49.78 6.79 1.69 10.16 2.24 3.92.65 4.6 2.79 3.81 6.32-3.44 15.3-5.78 30.81-8.11 46.31-3.65 24.18-6.45 48.46-8.91 72.79-4.2 41.51-6.29 83.1-6.6 124.8-.08 11.41.46 22.82.87 34.23a8.85 8.85 0 0 1-2.65 7.15c-11.36 11-25.27 16.54-40.72 18.65-2.9.4-4.31-.74-4.2-3.85.15-3.92 0-7.86 0-12.06-6.17 1-11.67 2.34-17.22 2.66-10.71.61-10.63.36-11.08-10.17-.68-16-.44-32.08.21-48.12.06-1.41 0-2.82 0-4.14-2-.34-1.83.88-2.09 1.77-5.74 19.47-11.57 38.92-15.07 59-.61 3.51-2.15 5.06-5.79 5.23-11.94.57-23.52-1.21-32.61-9.15a35.48 35.48 0 0 0-14.19-7.72c-12.55-3.63-20.08-11.51-20.56-25.2-.14-3.86-1.11-7.7-1.74-11.86-3.21 1.53-2 4-2.44 5.81a503.14 503.14 0 0 0-8.85 51.28c-.3 2.74-1.4 4.18-4.07 4.75-13 2.75-24.31.05-33.36-10a9.57 9.57 0 0 0-5.61-3.49c-12-2.37-19-7.16-26.64-18.48-.72 5.7-1.35 10.41-1.88 15.12-1.29 11.31-2.71 22.61-3.67 33.95-.31 3.68-2.25 5.05-5.2 5.37-12.31 1.32-23.37-1.53-32.24-10.59a10.81 10.81 0 0 0-5.72-3.39A37.79 37.79 0 0 1 475.19 335a6.43 6.43 0 0 0-3.68-2.31c-6.47-1.25-11.49-4.88-15.76-9.7a6.77 6.77 0 0 0-4.12-2.55c-15.55-2.33-24.9-13-32.9-25-8.58-12.91-12.45-27.62-14.45-42.86-.26-1.94-.54-3.87-.8-5.79-2 0-1.92 1.44-2.3 2.41-10.52 27.21-24 52.71-44.65 73.7-16.29 16.54-35.11 28.24-59.53 27-7.66-.4-14.72-2.72-18.43-10.44-1.11-2.3-3.23-1.89-5.14-2.17-10-1.49-18.49-5.26-23-15-2.61-5.61-2.52-11.52-1.75-17.5 6.46-50.07 14.51-99.9 22.57-149.72C277.9 114.02 284.7 73 292.83 32.21c.34-1.69.59-3.11 2.34-4 18.47-9.39 37.6-15.4 58.7-12.23 21.38 3.21 34.77 16.63 44 34.88 3.72 7.37 7.46 14.64 11.64 21.77C418.16 87.4 421 104 422.09 121q.07 1.23.21 2.46c.02.05.26.12.98.54ZM673.85 10.89c-1.61-1.23-3.34-1.1-5-1-17.48.65-33.55 5.64-47.72 16.05a27.81 27.81 0 0 0-10.54 15.75c-22.27 83.28-41.88 167.14-56.19 252.2-1.09 6.45.74 11.36 5.48 15.46a30.35 30.35 0 0 0 19.54 7.52c9.78.28 11.18-1 12.5-10.58.09-.66.19-1.31.3-2 8.38-52.17 21.55-103.3 33.46-154.7.79-3.4 1.9-6.72 4.22-9.42 1.16-1.36 2.45-3.07 4.53-2.27 1.74.66 2 2.52 2.09 4.21.06 1.15 0 2.31 0 3.47-.53 15.87-.84 31.76-1.65 47.62-1.63 31.9-3.9 63.75 2.23 95.52 1.4 7.2 4.47 11.77 11 14.44a65.76 65.76 0 0 0 22.12 5.14c4.89.22 8-1.06 7.93-6.66 0-2.1.78-4.21 1.23-6.32 5.64-26.42 14.44-51.94 22.39-77.71 8.37-27.14 16.6-54.32 25-81.45 1.14-3.68 2.28-8.77 7.38-7.75s3.43 6.07 2.93 9.61c-5.25 37.4-7.83 75-9.49 112.72-.92 20.82-1.6 41.67-.86 62.53.1 2.74.29 5 4.08 4.33 13.43-2.51 25.65-7.62 36.14-16.53 2.57-2.18 3.5-4.87 3.39-8.33-2.92-92.67 5.54-184.42 22.68-275.44.95-5-.52-6.47-4.84-7.19-20.84-3.47-39.82.69-56.67 13.62a23.61 23.61 0 0 0-8.17 11C711.78 68 700.1 101.28 688.61 134.6c-3.55 10.31-6.53 20.83-9.85 31.23-.71 2.2-1.68 4.9-4.49 4.49-3.18-.47-1.81-3.46-1.76-5.33.61-22.83.35-45.67.07-68.5-.34-27.79-1.22-55.6 1.48-83.34a10.72 10.72 0 0 0-.21-2.26ZM400.22 113.63c.06-16.14-1.25-32.12-6.69-47.44-6.19-17.42-15.55-32.6-34.19-38.77-19.08-6.32-37.23-1.65-54.69 6.91-2.38 1.17-2.48 3.36-2.86 5.42-3.39 18.21-7 36.39-10 54.66-7 42.08-14 84.18-20.6 126.33-4.56 29.08-9.26 58.15-12.49 87.42-1.17 10.58 3.9 17 14.47 18.89 11 1.91 21.47-.7 31.44-5 17-7.33 30.27-19.46 41.49-33.89 20.34-26.16 33.36-55.77 41.21-87.84 6.95-28.52 12.58-57.2 12.91-86.69Zm104.1 75.93c.16.57.39.93.31 1.19-6.29 21.94-12.67 43.85-16.73 66.34-.36 2-1.74 2.62-3.2 3.4-13.4 7.17-26.4 3.08-33.54-10.4-3.63-6.86-3.91-14.26-4.22-21.71-1.59-37.77 6.58-72.93 28-104.53 14.43-21.28 33-38.32 54.54-52.12 6.84-4.38 10.24-10.47 11.06-18.36a58.68 58.68 0 0 0-4.92-30.51c-6.14-13.61-5.82-11.4-16.1-5.3-16 9.49-28.47 23-40.55 36.94-42.46 49-63.54 106.22-66 170.57-.82 21.55 1.11 43 12.42 62 14.29 24 24.25 26.44 46.93 22.55 2.18-.37 4.28-1.77 6.39-1v11a12.59 12.59 0 0 0 5.05 10.32 38.88 38.88 0 0 0 27.33 8.47c3.48-.22 5.27-1.46 5.55-5.35.58-8.25 1.79-16.44 2.6-24.67 5.3-53.41 14.29-106.16 28.78-157.89 1.15-4.09-.14-5.9-3.77-6.91-2.23-.62-4.44-1.6-6.7-1.78-24.23-1.94-54.89 6.09-64.55 39.19-1.32 4.52-3 9-2.28 14.39 11.28-2.39 22.17-6.09 33.6-5.83Zm-13.94 11.64a125 125 0 0 0-16.15 4.18 3.35 3.35 0 0 0-2.54 3.08 204.34 204.34 0 0 0-2.64 38.43c.07 2.22.52 4.43.77 6.52 8.36.44 8.61.24 10-7 2.49-12.77 5.91-25.31 9.4-37.84.66-2.29 2.15-4.51 1.16-7.37Z"
+          className="cls-1"
+        />
+        <path
+          d="M301.1 270.71c1-8.33 1.86-16.32 2.92-24.29 6.89-51.87 16-103.33 28.38-154.19 2.07-8.5 11.55-11.79 18.07-5.79 8.06 7.42 10.89 17.31 11 27.94.15 21.89-.92 43.66-6.09 65.08-6.28 26-14.36 51.3-28.71 74.18-4 6.41-8.83 12-15.85 15.29a18.33 18.33 0 0 1-9.72 1.78ZM351 131.25c-6.21 26.28-12 52.64-15.67 79.41A265.3 265.3 0 0 0 351 131.25Z"
+          className="cls-1"
+        />
+        <path
+          d="M69 248.27H20c-10.13 0-17.14-4.66-19.35-12.94a29.73 29.73 0 0 1-.65-7.39V119.82c0-12.57 6.72-19.3 19.29-19.31h148.5c13.66 0 18.72 4.26 21.14 17.71a27.33 27.33 0 0 0 1.19 3c3.91-4.41 7.11-8.31 10.61-11.92s6.49-3.77 10.26-.21q6.48 6.14 12.61 12.64c4.29 4.55 4.21 6.39-.19 10.83-10.33 10.44-20.66 20.94-31.17 31.24a9.79 9.79 0 0 0-3.34 7.75c.14 18.35.08 36.7.06 55 0 14.73-6.71 21.54-21.52 21.63-18 .11-36.09.13-54.13-.07-3.94 0-7.2.57-9.5 3.91-7.93 11.51-21.06 11.83-32.78 15-3.65 1-6-2.13-5.15-6.34.8-3.94 1.99-7.88 3.12-12.41Zm25.66-142.56H19.2c-9.1 0-13.91 4.53-13.92 13.11q-.09 55.29 0 110.57c0 8.64 4.63 13 13.29 13.05 14.4 0 28.8-.05 43.2 0 4.59 0 9.25.79 10.67-5.49.28-1.24 1.94-2.18 3-3.23l104.87-104.79c1.52-1.52 2.94-2.91 3.14-5.37 1-12.6-3.67-17.83-16.3-17.84q-36.24-.03-72.49-.01Zm-23.54 156c6.57-1.7 12.08-3.22 17.64-4.54a13.78 13.78 0 0 0 6.75-3.88q42.81-42.84 85.64-85.64c1.1-1.11 2.81-1.9 2.79-2.61l-5.89-7.41c-2.24.54-3.06 1.88-4.09 2.91q-25.15 25-50.26 50.13-17.92 17.9-35.86 35.76c-1.59 1.58-3.49 4-5.82 1.87-2.53-2.3.11-4.06 1.61-5.72.33-.37.69-.72 1-1.07q41.64-41.55 83.26-83.13c1.71-1.7 3.92-3.05 4.76-5.53-1.11-1.55-2.11-3.05-4-3.69a7.15 7.15 0 0 0-1.28.79Q122.8 194.44 78.21 239a7 7 0 0 0-2 2.8c-1.66 6.2-3.21 12.42-5.09 19.86Zm44.43-19.15h55.07c8.16 0 12.74-4.71 12.76-13q.06-25.53 0-51.06c0-1.2.59-2.67-.55-3.37Zm80.51-120.6 15.2 15.36a83.53 83.53 0 0 0 9.53-9.56l-15.27-15.25a80.41 80.41 0 0 0-9.46 9.4Zm4.65 13.93L182.35 154c1.62 1.77 3.47 4.08 5.52 5.33l18.33-18.05c-1.9-1.89-3.99-3.96-5.49-5.44Zm-5.87-5.31c-.82-1.42-1.75-3.13-3.43-3.24L174 145.64a5.45 5.45 0 0 0 3.22 2.7Z"
+          className="cls-2"
+        />
+        <path
+          d="M42.16 146.2c-6.46-2.86-11.55-5.09-16.63-7.35-1.66-.74-3.28-1.58-4.94-2.31a3.57 3.57 0 0 1-2.13-4.87c.79-2.32 2.76-2.66 4.87-2.25 3.31.63 6.06 2.59 9.07 3.93 6.18 2.76 12.35 5.55 18.55 8.29 2.12.94 3.7 2.29 3.68 4.77s-1.78 3.27-3.68 4.13c-8.42 3.83-16.78 7.8-25.22 11.58-2.55 1.14-5.72 2.51-7.18-1.22-1.36-3.47 1.41-4.89 4.18-6.08 6.02-2.59 11.99-5.31 19.43-8.62Z"
+          className="cls-2"
+        />
+      </g>
+    </g>
+  </svg>
+);
+export default SvgComponent;
