+++ src/components/atoms/logo/types.ts
@@ -1,0 +1,5 @@
+export interface LogoProps {
+  width?: number;
+  height?: number;
+  className?: string;
+}
