+++ src/components/atoms/seo/types.ts
@@ -1,0 +1,3 @@
+export interface SeoProps {
+  pageTitle: string;
+}
