+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './atoms';
+export * from './molecules';
+export * from './organisms';
