+++ src/components/molecules/mode/index.ts
@@ -1,0 +1,2 @@
+export { default as DarkMode } from './DarkMode';
+export { default as DarkModeMobile } from './DarkModeMobile';
