+++ src/components/organisms/index.ts
@@ -1,0 +1,2 @@
+export { default as Hero } from './hero';
+export { default as Social } from './social';
