--- src/data/HeaderMenu.js
@@ -1,44 +1,0 @@
-import { AiOutlineHome } from 'react-icons/ai';
-import { FaBlogger, FaRegUser } from 'react-icons/fa';
-import { CgNotes } from 'react-icons/cg';
-import { FiCodesandbox } from 'react-icons/fi';
-import { RiContactsBookLine } from 'react-icons/ri';
-
-export default [
-  {
-    id: '01',
-    name: 'Home',
-    routePath: '/',
-    icon: <AiOutlineHome />,
-  },
-  // {
-  //   id: "02",
-  //   name: "About",
-  //   routePath: "/about",
-  //   icon: <FaRegUser />,
-  // },
-  // {
-  //   id: "06",
-  //   name: "Resume",
-  //   routePath: "/resume",
-  //   icon: <CgNotes />,
-  // },
-  // {
-  //   id: "03",
-  //   name: "Works",
-  //   routePath: "/works",
-  //   icon: <FiCodesandbox />,
-  // },
-  // {
-  //   id: "04",
-  //   name: "Blogs",
-  //   routePath: "/blogs",
-  //   icon: <FaBlogger />,
-  // },
-  // {
-  //   id: "05",
-  //   name: "Contact",
-  //   routePath: "/contact",
-  //   icon: <RiContactsBookLine />,
-  // },
-];
