--- src/data/addressData.js
@@ -1,59 +1,0 @@
-module.exports = [
-  {
-    id: '01',
-    icon: '/images/contact/phone-call 1.png',
-    title: 'Phone ',
-    info: (
-      <>
-        <p>
-          <a href="tel:+452 666 386" className="text-gray-lite text-lg dark:text-[#A6A6A6] ">
-            +452 666 386
-          </a>
-        </p>
-
-        <p>
-          <a href="tel:+654 764 878" className="text-gray-lite text-lg dark:text-[#A6A6A6] ">
-            +654 764 878
-          </a>
-        </p>
-      </>
-    ),
-    bg: '#FCF4FF',
-  },
-  {
-    id: '02',
-    icon: '/images/contact/email 1.png',
-    title: 'Email ',
-    info: (
-      <>
-        <p>
-          <a
-            href="mailto:ibthemes21@gmail.com"
-            className="text-gray-lite text-lg dark:text-[#A6A6A6] "
-          >
-            ibthemes@gmail.com
-          </a>
-        </p>
-
-        <p>
-          <a href="mailto:xyz@abc.com" className="text-gray-lite text-lg dark:text-[#A6A6A6] ">
-            mailto:xyz@abc.com
-          </a>
-        </p>
-      </>
-    ),
-    bg: '#EEFBFF',
-  },
-  {
-    id: '03',
-    icon: '/images/contact/map 1.png',
-    title: 'Address ',
-    info: (
-      <>
-        <p className="text-gray-lite text-lg dark:text-[#A6A6A6] ">Maount View, Oval</p>
-        <p className="text-gray-lite text-lg dark:text-[#A6A6A6] ">Road, New York, USA</p>
-      </>
-    ),
-    bg: '#F2F4FF',
-  },
-];
