--- src/data/brandSlideData.js
@@ -1,26 +1,0 @@
-module.exports = [
-  {
-    id: 1,
-    img: '/images/slider/brand-1.png',
-  },
-  {
-    id: 2,
-    img: '/images/slider/brand-2.png',
-  },
-  {
-    id: 3,
-    img: '/images/slider/brand-3.png',
-  },
-  {
-    id: 4,
-    img: '/images/slider/brand-4.png',
-  },
-  {
-    id: 5,
-    img: '/images/slider/brand-5.png',
-  },
-  {
-    id: 6,
-    img: '/images/slider/brand-1.png',
-  },
-];
