--- src/data/serviceData.js
@@ -1,50 +1,0 @@
-module.exports = [
-  {
-    id: '1',
-    icon: '/images/icons/icon-1.svg',
-    title: 'Ui/Ux Design',
-    des: 'Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam euismod volutpat.',
-    color: '#D566FF',
-    bg: '#FCF4FF',
-  },
-  {
-    id: '2',
-    icon: '/images/icons/icon-2.svg',
-    title: 'App Development',
-    des: 'Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam euismod volutpat.',
-    color: '#DDA10C',
-    bg: '#FEFAF0',
-  },
-  {
-    id: '3',
-    icon: '/images/icons/icon-3.svg',
-    title: 'Photography',
-    des: 'Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam euismod volutpat.',
-    color: '#8774FF',
-    bg: '#FCF4FF',
-  },
-  {
-    id: '4',
-    icon: '/images/icons/icon-4.svg',
-    title: 'Photography',
-    des: 'Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam euismod volutpat.',
-    color: '#FF6080',
-    bg: '#FFF4F4',
-  },
-  {
-    id: '5',
-    icon: '/images/icons/icon-5.svg',
-    title: 'Managment',
-    des: 'Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam euismod volutpat.',
-    color: '#FF75D8',
-    bg: '#FFF0F8',
-  },
-  {
-    id: '6',
-    icon: '/images/icons/icon-6.svg',
-    title: 'Web Development',
-    des: 'Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam euismod volutpat.',
-    color: '#269FFF',
-    bg: '#F3FAFF',
-  },
-];
