--- src/data/worksData.js
@@ -1,128 +1,0 @@
-module.exports = [
-  {
-    id: '1',
-    tag: 'UI/UX',
-    title: 'Chul urina',
-    img: '/images/work_images/1.jpg',
-    imgSmall: '/images/work_images/small/1.jpg',
-    bg: '#FFF0F0',
-    client: 'Envato',
-    langages: 'Photoshop, Figma',
-    link: 'https://www.envato.com',
-    linkText: 'www.envato.com',
-    description:
-      'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Mollitia placeat magnam possimus iusto blanditiis pariatur labore explicabo quo repellat hic dolorum numquam asperiores, voluptatum fugiat reiciendis aspernatur, non, odio aperiam voluptas ex tempora vitae. Dolor, consequatur quidem! Quas magni distinctio dolorum dolore natus, vel numquam accusamus. Nostrum eligendi recusandae qui tempore deserunt!',
-  },
-  {
-    id: '2',
-    tag: 'Web Design',
-    title: 'Aura Dione',
-    img: '/images/work_images/2.jpg',
-    imgSmall: '/images/work_images/small/2.jpg',
-    bg: '#FFF3FC',
-    client: 'Themeforest',
-    langages: 'HTML, CSS, Javascript',
-    link: 'https://www.themeforest.net',
-    linkText: 'themeforest.net',
-    description:
-      '  Lorem ipsum dolor sit amet consectetur, adipisicing elit. Cupiditate non suscipit voluptatibus minima ullam maiores sequi nihil placeat error, vero eaque doloremque reiciendis amet pariatur consequuntur. Nostrum, dolore, amet eligendi ipsam enim quisquam, corrupti asperiores nihil excepturi aspernatur placeat iure.',
-  },
-  {
-    id: '3',
-    tag: 'Logo',
-    title: 'Chul urina',
-    img: '/images/work_images/3.jpg',
-    imgSmall: '/images/work_images/small/3.jpg',
-    bg: '#FFF0F0',
-    client: 'Freepik',
-    langages: 'Illustrator',
-    link: 'https://www.freepik.com/free-photos-vectors/market-logo',
-    linkText: 'www.freepik.com',
-    description:
-      '  Lorem ipsum dolor sit amet consectetur, adipisicing elit. Cupiditate non suscipit voluptatibus minima ullam maiores sequi nihil placeat error, vero eaque doloremque reiciendis amet pariatur consequuntur. Nostrum, dolore, amet eligendi ipsam enim quisquam, corrupti asperiores nihil excepturi aspernatur placeat iure.',
-  },
-  {
-    id: '4',
-    tag: 'Video',
-    title: 'Chul urina',
-    img: '/images/work_images/4.jpg',
-    imgSmall: '/images/work_images/small/4.jpg',
-    bg: '#E9FAFF',
-    client: 'Envato',
-    langages: 'After Effect',
-    link: 'https://www.envato.com',
-    linkText: 'www.envato.com',
-    description:
-      '  Lorem ipsum dolor sit amet consectetur, adipisicing elit. Cupiditate non suscipit voluptatibus minima ullam maiores sequi nihil placeat error, vero eaque doloremque reiciendis amet pariatur consequuntur. Nostrum, dolore, amet eligendi ipsam enim quisquam, corrupti asperiores nihil excepturi aspernatur placeat iure.',
-  },
-  {
-    id: '5',
-    tag: 'UI/UX',
-    title: 'Chul urina',
-    img: '/images/work_images/5.jpg',
-    imgSmall: '/images/work_images/small/5.jpg',
-    bg: '#FFFAE9',
-    client: 'Envato',
-    langages: 'Photoshop',
-    link: 'https://www.envato.com',
-    linkText: 'www.envato.com',
-    description:
-      '  Lorem ipsum dolor sit amet consectetur, adipisicing elit. Cupiditate non suscipit voluptatibus minima ullam maiores sequi nihil placeat error, vero eaque doloremque reiciendis amet pariatur consequuntur. Nostrum, dolore, amet eligendi ipsam enim quisquam, corrupti asperiores nihil excepturi aspernatur placeat iure.',
-  },
-  {
-    id: '6',
-    tag: 'Video',
-    title: 'Chul urina',
-    img: '/images/work_images/6.jpg',
-    imgSmall: '/images/work_images/small/6.jpg',
-    bg: '#F4F4FF',
-    client: 'Envato',
-    langages: 'Vimeo',
-    link: 'https://www.envato.com',
-    linkText: 'www.envato.com',
-    description:
-      '  Lorem ipsum dolor sit amet consectetur, adipisicing elit. Cupiditate non suscipit voluptatibus minima ullam maiores sequi nihil placeat error, vero eaque doloremque reiciendis amet pariatur consequuntur. Nostrum, dolore, amet eligendi ipsam enim quisquam, corrupti asperiores nihil excepturi aspernatur placeat iure.',
-  },
-  {
-    id: '7',
-    tag: 'UI/UX',
-    title: 'Chul urina',
-    img: '/images/work_images/7.jpg',
-    imgSmall: '/images/work_images/small/7.jpg',
-    bg: '#FFF0F8',
-    client: 'Envato',
-    langages: 'Photoshop',
-    link: 'https://www.envato.com',
-    linkText: 'www.envato.com',
-    description:
-      '  Lorem ipsum dolor sit amet consectetur, adipisicing elit. Cupiditate non suscipit voluptatibus minima ullam maiores sequi nihil placeat error, vero eaque doloremque reiciendis amet pariatur consequuntur. Nostrum, dolore, amet eligendi ipsam enim quisquam, corrupti asperiores nihil excepturi aspernatur placeat iure.',
-  },
-  {
-    id: '8',
-    tag: 'Web Design',
-    title: 'Chul urina',
-    img: '/images/work_images/8.jpg',
-    imgSmall: '/images/work_images/small/8.jpg',
-    bg: '#FFF0F8',
-    client: 'Envato',
-    langages: 'HTML, CSS, Javascript',
-    link: 'https://www.envato.com',
-    linkText: 'www.envato.com',
-    description:
-      '  Lorem ipsum dolor sit amet consectetur, adipisicing elit. Cupiditate non suscipit voluptatibus minima ullam maiores sequi nihil placeat error, vero eaque doloremque reiciendis amet pariatur consequuntur. Nostrum, dolore, amet eligendi ipsam enim quisquam, corrupti asperiores nihil excepturi aspernatur placeat iure.',
-  },
-  {
-    id: '9',
-    tag: 'Logo',
-    title: 'Chul urina',
-    img: '/images/work_images/9.jpg',
-    imgSmall: '/images/work_images/small/9.jpg',
-    bg: '#FCF4FF',
-    client: 'Feepik',
-    langages: 'Figma',
-    link: 'https://www.freepik.com/free-photos-vectors/market-logo',
-    linkText: 'www.freepik.com',
-    description:
-      '  Lorem ipsum dolor sit amet consectetur, adipisicing elit. Cupiditate non suscipit voluptatibus minima ullam maiores sequi nihil placeat error, vero eaque doloremque reiciendis amet pariatur consequuntur. Nostrum, dolore, amet eligendi ipsam enim quisquam, corrupti asperiores nihil excepturi aspernatur placeat iure.',
-  },
-];
